Example dataset for ForestData: Paracou, plot 6

A built-in example dataset used in ForestData's examples and vignette. It corresponds to the squares 1 and 3 of the 6th Plot of the Paracou Experiment database

This dataset was obtain from the Paracou database and exclusively aims at illustrating ForestData features. It has been purposedly modified: plot 6 square 1 is provided as-is, but plot 6 square 3 misses several years to simulate two plots with different censusing resolution.

example_census

data(example_census)

Format

A data frame with 25533 rows and 12 variables:

Forest

Character, general information: Forest sector where the plot to which trees belong is located. Here, always Paracou:P6

Plot

Integer (could be character), the index of the plot to which trees belong

idTree

Integer (could be character), the unique id of the tree

Family

Character, taxonomic information : botanical family

Genus

Character, taxonomic information : botanical genus

Species

Character, taxonomic information : species name

CensusYear

Integer, the year at which the tree is censused

CodeAlive

Logical or integer, tree life status: 1 for alive, 0 for dead. NA, if any, indicate that the tree was not seen for the corresponding census

CodeMeas

Character, not important here. Additionan informations relative to measurements

Circ

Numeric. Tree size expressed as stem circumference, in centimeters

binomial_name

Taxonomic information: full binomial species name

POM

Numeric. Point of Measurement, in meters.

An object of class "data.frame" with

ForestThe location of the plots included in the dataset

PlotThe plot indices correspondinc to the plot a given tree belongs to - In this dataset, the plots all correspond to subplots of Paracou's Plot6 (control, undisturbed)

idTreeUnique tree ID

FamilyTaxonomical information, the plant family of each tree

GenusTaxonomical information, the genus of each tree

SpeciesTaxonomical information, the species name of each tree

binomial_nameTaxonomical information, the complete binomial name (genus_species) of each tree

CensusYearThe census year at which each data have been acquired

CodeAliveWhether the tree is dead (0) or alive (1). NAs do not exist in the original example dataset but can be found after using correct_alive (see the documentation)

CircThe circumference of the tree

POMThe Point of Measurement (in meters) for each data

Source

https://paracou.cirad.fr/

On demand

References

The Paracou Research Station and Forest sDisturbance Experiment (Paracou Research Station)

Examples

data(example_census)