Example dataset for ForestData: Paracou, plot 6, with corrected status

The same built-in example dataset as example_census, corrected with correct_alive

example_status_corr

Format

A data frame with 25634 rows and 13 variables:

Forest

Character, Forest sector where the plot is located

Plot

Integer (could be character), the index of the plot to which trees belong

idTree

Integer (could be character), the unique id of the tree

Family

Character, taxonomic information : botanical family

Genus

Character, taxonomic information : botanical genus

Species

Character, taxonomic information : species name

CensusYear

Integer, the year at which the tree is censused

CodeAlive

Logical or integer, tree life status: 1 for alive, 0 for dead. NA, if any, indicate that the tree was not seen for the corresponding census

CodeMeas

Character, not important here. Additionan informations relative to measurements

Circ

Numeric. Tree size expressed as stem circumference, in centimeters

binomial_name

Taxonomic information: full binomial species name

POM

Numeric. Point of Measurement, in meters.

status_corr

Integer, interpretable as logical. Corrected vital status obtained with correct_alive

Source

https://paracou.cirad.fr/