Download forest canopy height

Download the ETH Global Sentinel-2 10m Canopy Height (2020)

Usage

fd_canopy_height(
  x = NULL,
  lon = NULL,
  lat = NULL,
  layer = "chm",
  crop = FALSE,
  ...
)

Arguments

x

An sf or SpatVector object. It will retrieve the necessary tiles to cover the area (if lat and lon are specified, this argument is ignored)

lon

A number specifying the longitude of the area where we want the tile

lat

A number specifying the latitude of the area where we want the tile

layer

A string for the layer to download. The default "chm" downloads the Canopy Height Model, while "std" downloads the standard deviation. If you want both layers, use "all"

crop

When x is specified, whether to crop the tiles(s) to the object

...

additional arguments passed to the `terra::crop()`

Value

A SpatRaster

Details

Data may be freely used for research, study, or teaching, but be cited appropriately (see references below).

References

Lang, Nico, Walter Jetz, Konrad Schindler, and Jan Dirk Wegner. "A high-resolution canopy height model of the Earth." arXiv preprint arXiv:2204.08322 (2022).

Examples

if (FALSE) { # \dontrun{
fd_canopy_height(lon = -7.27, lat = 42.43)
} # }