EU-Trees4F Database

Download data for tree species distribution in Europe for current (2005) distribution, and future distribution (2035, 2065, 2095).

Usage

fd_forest_eutrees4f(
  species,
  model = "clim",
  period = "all",
  scenario = "rcp45",
  type = "bin",
  distrib = "pot",
  quiet = FALSE
)

Arguments

species

a character vector of length 1 with the Latin name of the tree species (genus and species)

model

a character vector of length 1 with the name of the ensemble projection. One of 'clim' or 'sdms' (see details)

period

a numeric or character vector of length 1 with the center of the 30-year time period used for the model. One of '2005', '2035', '2065', '2095', or 'all' (see details)

scenario

a character vector of length 1 with the climate change scenario used. One of 'rcp45' or 'rcp85' (see details)

type

a character vector of length 1 with the type of output layer. One of 'bin', 'prob' or 'std' (see details)

distrib

a character vector of length 1 with the type of distribution. One of 'nat', 'pot', 'disp' or 'disp_lu' (see details)

quiet

if TRUE, suppress any message or progress bar

Value

A single-band or multi-band SpatRaster

Details

Data may be freely used for research, study, or teaching, but be cited appropriately (see references below).

The data of EU-Trees4F database represent the distribution of the main woody species in Europe at 5 arc-minutes (~ 10 km) spatial resolution, in the Lambert Azimuthal Equal Area (EPSG:3035) CRS. The possible models to download are the following:

Model: type of model used

clim: climatic ensemble. A ensemble mean model that projects a consensus model from biomod2 into future conditions using the average of 11 Regional Climate Models (RCM).

sdms: Species Distribution Model (SDM) ensemble. A model that projects the consensus model for every single RCM, and then it averages the output of then 11 SDMs.

Period: 30-year time period

2005: for current projections. This option ignores the scenario argument. They are not available for model = 'sdms' with type = 'std'.

2035: average of 2020-2050

2065: average of 2050-2080

2095: average of 2080-2110

all: get the four periods (or three for std type). Note that for some species or configurations this might fail, because the raster extent might not match in different periods

Scenario: climate change scenario

rcp45: a climate change scenario that assumes moderate emissions reductions

rcp85: a climate change scenario with high greenhouse gas emissions and limited mitigation efforts

Type: type of output layer

bin: binary distribution map, where 1 represents presence of the tree species, while 0 represents absence of the tree species, derived from the prob map

prob: probability distribution map (0-1000). Represents the probability of being the potential distribution of the species

std: standard deviation of prob map. Only available for model = 'sdms'.

Distrib: type of species distribution

nat: realized distribution (masked with native range). Only available with type = 'bin'

pot: potential distribution

disp: natural dispersal model (migclim). Only available with type = 'bin'

disp_lu: natural dispersal model clipped by forest areas. Only available with type = 'bin'

References

Mauri, Achille; Cescatti, Alessandro; GIRARDELLO, MARCO; Strona, Giovanni; Beck, Pieter; Caudullo, Giovanni; et al. (2022). EU-Trees4F. A dataset on the future distribution of European tree species.. figshare. Collection. https://doi.org/10.6084/m9.figshare.c.5525688.v2

See also

Examples

# \donttest{
# Download data for Betula pendula
betula_pendula_sr <- fd_forest_eutrees4f(species = "Betula pendula")
#> Cite this dataset using <https://doi.org/10.6084/m9.figshare.c.5525688.v2>
# }