Forest Extent

This function is deprecated in favour of fd_forest_glad. Download the Forest Extent raster from the Global Land Analysis & Discovery by using a vectorial object or a pair of coordinates (latitude, longitude).

Usage

fd_forest_extent_glad(
  x = NULL,
  lon = NULL,
  lat = NULL,
  year = 2020,
  crop = FALSE,
  quiet = FALSE,
  ...
)

Arguments

x

a sf or SpatVector object. It will retrieve the necessary tiles to cover the area (if lat and lon are specified, this argument is ignored)

lon

a number specifying the longitude of the area where we want the tile

lat

a number specifying the latitude of the area where we want the tile

year

year of the forest extent data. One of 2000, 2020 or 'all'

crop

when x is specified, whether to crop the tiles(s) to the object

quiet

if TRUE, suppress any message or progress bar

...

additional arguments passed to the crop function

Value

SpatRaster object

Details

The Forest Extent Map is a product offered by the Global Land Analysis & Discovery organization. The spatial resolution of the product is 0.00025º (approximately 30 meters at the Equator), and it's distributed in tiles of 10ºx10º. Pixels with forest height > 5 meters are classified as the forest class.

Note that each tile is stored as a raster file of 1.5 GB, so for big extensions the function might take some time to retrieve the data.

References

Potapov P., Hansen M.C., Pickens A., Hernandez-Serna A., Tyukavina A., Turubanova S., Zalles V., Li X., Khan A., Stolle F., Harris N., Song X.-P., Baggett A., Kommareddy I., Kommareddy A. (2022) The global 2000-2020 land cover and land use change dataset derived from the Landsat archive: first results. Frontiers in Remote Sensing doi:10.3389/frsen.2022.856903

Examples

# \donttest{
 # Get tile for Galicia (Spain)
 galicia_forest_extent <- fd_forest_extent_glad(lon = -7.8, lat = 42.7, year = 2020)
#> Warning: `fd_forest_extent_glad()` was deprecated in forestdata 0.2.0.
#> ℹ Please use the `model` argument of `fd_forest_glad()` instead.
#> 1 tile(s) were found. A total of 1.5 GB of data will be read into R. This may take a while.
#> Cite this dataset using <https://doi.org/10.3389/frsen.2022.856903>
# }