Global Land Analysis & Discovery datasets

Download data from GLAD database including forest extent, forest height, and land cover at ~30m spatial resolution

Usage

fd_forest_glad(
  x = NULL,
  lon = NULL,
  lat = NULL,
  model = "extent",
  year = 2020,
  crop = FALSE,
  mask = FALSE,
  merge = FALSE,
  quiet = FALSE
)

Arguments

x

a sf or SpatVector object. It will retrieve the necessary tiles to cover the area (if lat and lon are specified, this argument is ignored)

lon

a number specifying the longitude of the area where we want the tile

lat

a number specifying the latitude of the area where we want the tile

model

a character vector of length 1 indicating the model to retrieve (see details)

year

year of the data (see details)

crop

when x is specified, whether to crop the tile(s) to the object

mask

when x is specified, whether to mask the tile(s) to the object

merge

if FALSE (default), it will merge the tiles into one raster. If FALSE a SpatRasterCollection will be returned.

quiet

if TRUE, suppress any message or progress bar

Value

SpatRaster object

Details

The Global Land Analysis & Discovery (GLAD) includes several datasets which can be accessed through the model argument:

landcover: global land cover and land use dataset. Dataset divided into 10ºx10º tiles containing measures of bare ground and tree height inside and outside of wetlands, seasonal water percent, binary labels of built-up, permanent ice/snow, and cropland. Available for the years 2000, 2005, 2010, 2015, and 2020.

landcover-change: changes of landcover from 2000 to 2020. Argument year is ignored.

extent: dataset showing presence of forest, defined as wildland, managed, and planted tree cover including agroforestry and orchards. Includes areas where the vegetation is taller than 5 meters. Available for the years 2000 and 2020.

height: dataset measuring the height of woody vegetation taller than 3 meters. Available for the years 2000 and 2020.

The spatial resolution of the product is 0.00025º (approximately 30 meters at the Equator), and it's distributed in tiles of 10ºx10º.

Note that each tile is stored as a raster file of 1.5 GB, so for big extensions the function might take some time to retrieve the data.

References

Potapov P., Hansen M.C., Pickens A., Hernandez-Serna A., Tyukavina A., Turubanova S., Zalles V., Li X., Khan A., Stolle F., Harris N., Song X.-P., Baggett A., Kommareddy I., Kommareddy A. (2022) The global 2000-2020 land cover and land use change dataset derived from the Landsat archive: first results. Frontiers in Remote Sensing doi:10.3389/frsen.2022.856903

P. Potapov, X. Li, A. Hernandez-Serna, A. Tyukavina, M.C. Hansen, A. Kommareddy, A. Pickens, S. Turubanova, H. Tang, C.E. Silva, J. Armston, R. Dubayah, J. B. Blair, M. Hofton (2020) Mapping and monitoring global forest canopy height through integration of GEDI and Landsat data. Remote Sensing of Environment, 112165.doi:10.1016/j.rse.2020.112165

Examples

# \donttest{
 # Get tile for Galicia (Spain)
 galicia_forest_extent <- fd_forest_glad(lon = -7.8, lat = 42.7, year = 2020)
#> ℹ 1 tile(s) were found. A total of 1.5 GB of data will be read into R. This may take a while.
#> ✔ Cite this dataset using https://doi.org/10.1016/j.rse.2020.112165
# }