Spanish Forest Inventory

Download the tables and SIG data from the Spanish Forest Inventory

Usage

fd_inventory_spain(
  province,
  ifn = 4,
  database = "field",
  path_metadata = NULL,
  quiet = FALSE
)

Arguments

province

a character string of length 1 with the name of a Spanish province

ifn

number of Spanish Forest Inventory (from 2 to 4)

database

the name of the database (either 'field' or 'gis')

path_metadata

a character string of length 1 with the path to store the metadata of the selected database. The default path_metadata = NULL does not download the metadata

quiet

if TRUE, suppress any message or progress bar

Value

A list with the tables

Details

The IFN2 doesn't have 'gis' data for Asturias, Cantabria and Navarra.

In the future a function to process the data will be added.

References

https://www.miteco.gob.es/es/biodiversidad/temas/inventarios-nacionales/inventario-forestal-nacional.html

See also

Examples

# \donttest{
# Download MFE50 for Canary Islands
canarias_ifn4_lst <- fd_inventory_spain("Canarias")
#> Error in pkgdown::build_site_github_pages(new_process = FALSE, install = FALSE): ! ODBC failed with error 00000 from [unixODBC][Driver Manager].
#> ✖ Can't open lib 'Microsoft Access Driver (*.mdb, *.accdb)' : file not found
#> ℹ From nanodbc/nanodbc.cpp:1150.

cantabria_ifn3_gis_lst <- fd_inventory_spain("cantabria", ifn = 3, database = "gis")
#> Error in pkgdown::build_site_github_pages(new_process = FALSE, install = FALSE): ! ODBC failed with error 00000 from [unixODBC][Driver Manager].
#> ✖ Can't open lib 'Microsoft Access Driver (*.mdb, *.accdb)' : file not found
#> ℹ From nanodbc/nanodbc.cpp:1150.
# }