EU Forest Species database

Downloads the EU Forest Species database, an European database of more than 500,000 forest tree species occurrences

Usage

fd_occ_euforest(species = NULL, country = NULL, spatial = FALSE, quiet = FALSE)

Arguments

species

a character vector with the name of one or more tree species

country

a character vector with either ISO2 codes, ISO3 codes or full country names (not mixed) to filter out the data

spatial

logical. Whether to retrieve a tibble or a sf object

quiet

if TRUE, suppress any message or progress bar

Value

a tibble or a sf object

References

A high resolution pan-European tree occurrence dataset doi:10.6084/m9.figshare.c.3288407.v1

See also

Examples

# \donttest{
## Download full database as tibble
euforest_tbl <- fd_occ_euforest()
#> ℹ Downloading data...
#> ✔ Downloaded [3.6s]
#> 
#> ℹ Preparing data...
#> ✔ Prepared [9ms]
#> 
#> ✔ Cite this dataset using <https://doi.org/10.6084/m9.figshare.c.3288407.v1>

## Download full database as spatial
euforest_sf <- fd_occ_euforest(spatial = TRUE)
#> ℹ Downloading data...
#> ✔ Downloaded [6ms]
#> 
#> ℹ Preparing data...
#> ✔ Prepared [8ms]
#> 
#> ✔ Cite this dataset using <https://doi.org/10.6084/m9.figshare.c.3288407.v1>

## Download data for Abies alba for Czechia and Germany
euforest_cz_ge_sf <- fd_occ_euforest(species = "Abies alba", country = c("Czechia", "Germany"))
#> ℹ Downloading data...
#> ✔ Downloaded [6ms]
#> 
#> ℹ Preparing data...
#> ✔ Prepared [8ms]
#> 
#> ✔ Cite this dataset using <https://doi.org/10.6084/m9.figshare.c.3288407.v1>
# }