Compute Basal Area in Forest Inventories

compute_ba computes Basal Area according to user-specified grouping variables -Plot, species, etc- and time.

compute_ba(data, by = c("Plot", "binomial_name"),
  size_col = "size_corr", measure_type = getOption("measure_type"),
  time_col = getOption("time_col"), status_col = "status_corr",
  surface = 1.5625)

Arguments

data
A data.frame containing a time-series tree-wise forest inventory -i.e. every line is a single tree measurement for a single year.

by
A character vector containing the name of the columns containing the variables -other than census time- according to which the result will be aggregated. Be it plots, subplots or species name...

size_col
character containing the name of the tree size measurements column -either circumference or diameter.

measure_type
A single character indicating whether tree sizes are given in circumferences -"C"- or diameter -"D"-.

time_col
A single character containing the name of the column corresponding to census years

surface
Either a scalar containing the surface area of each plot -if they have the same dimensions- or a data.frame of the surface area according to some of the grouping variables -e.g. Plot and subplot. However, defaults to FALSE and in this case, only absolute BA is returned.

Value

A dataframe containing, for each combination of grouping variables, the plot-level and per-ha basal area for a given census year.

Examples

if (FALSE) {
data("example_census")
compute_ba(example_census,
size_col = "CircCorr",
measure_type = "C",
time_col = "CensusYear",
by = c("Species","Plot"),
surface = 1.5625)
}