Example Forest Census to Illustrate ForestData Functions

This dataset was obtain from the Paracou database and exclusively aims at illustrating ForestData features. It has been purposedly modified: plot 6 square 1 is provided as-is, but plot 6 square 3 misses several years to simulate two plots with different censusing resolution.

data(example_census)

Format

An object of class "data.frame" with

ForestThe location of the plots included in the dataset

PlotThe plot indices correspondinc to the plot a given tree belongs to - In this dataset, the plots all correspond to subplots of Paracou's Plot6 (control, undisturbed)

idTreeUnique tree ID

FamilyTaxonomical information, the plant family of each tree

GenusTaxonomical information, the genus of each tree

SpeciesTaxonomical information, the species name of each tree

binomial_nameTaxonomical information, the complete binomial name (genus_species) of each tree

CensusYearThe census year at which each data have been acquired

CodeAliveWhether the tree is dead (0) or alive (1). NAs do not exist in the original example dataset but can be found after using correct_alive (see the documentation)

CircThe circumference of the tree

POMThe Point of Measurement (in meters) for each data

Source

On demand

References

The Paracou Research Station and Forest sDisturbance Experiment (Paracou Research Station)

Examples

data(example_census)