Forest Canopy Height

Download the ETH Global Sentinel-2 10m Canopy Height (2020) or the Meta High Resolution 1m Global Canopy Height Map

Usage

fd_canopy_height(
  x = NULL,
  lon = NULL,
  lat = NULL,
  model = "eth",
  layer = "chm",
  crop = FALSE,
  mask = FALSE,
  merge = FALSE,
  quiet = FALSE
)

Arguments

x

a sf or SpatVector object. It will retrieve the necessary tiles to cover the area (if lat and lon are specified, this argument is ignored)

lon

a number specifying the longitude of the area where we want the tile

lat

a number specifying the latitude of the area where we want the tile

model

a string specifying the model to download. One of "eth" or "meta" (see details)

layer

a string for the layer to download (valid only for eth). The default "chm" downloads the Canopy Height Model, while "std" downloads the standard deviation. If you want both layers, use "all"

crop

when x is specified, whether to crop the tile(s) to the object

mask

when x is specified, whether to mask the tile(s) to the object

merge

if FALSE (default), it will merge the tiles into one raster. If FALSE a SpatRasterCollection will be returned.

quiet

if TRUE, suppress any message or progress bar

Value

A SpatRaster or SpatRasterCollection

Details

There are currently two global canopy height models available within this function.

eth: the ETH Global Sentinel-2 10m Canopy Height from the year 2020. Visit https://www.research-collection.ethz.ch/handle/20.500.11850/609802 for more information

meta: the Meta High Resolution 1m Global Canopy Height. Visit doi:10.1016/j.rse.2023.113888 for more information

Data may be freely used for research, study, or teaching, but be cited appropriately (see references below).

References

Lang, Nico, Walter Jetz, Konrad Schindler, and Jan Dirk Wegner. "A high-resolution canopy height model of the Earth." arXiv preprint arXiv:2204.08322 (2022).

Tolan, J., Yang, H.I., Nosarzewski, B., Couairon, G., Vo, H.V., Brandt, J., Spore, J., Majumdar, S., Haziza, D., Vamaraju, J. and Moutakanni, T., 2024. Very high resolution canopy height maps from RGB imagery using self-supervised vision transformer and convolutional decoder trained on aerial lidar. Remote Sensing of Environment, 300, p.113888.

Examples

if (FALSE) { # \dontrun{
## Get 10m resolution CHM
eth_model <- fd_canopy_height(lon = -7.27, lat = 42.43)

## Get 1m resolution CHM
meta_model <- fd_canopy_height(lon = -7.27, lat = 42.43, model = "meta")
} # }